Create a utility function for resetting states
